Introduce GPUTileSwizzleUtils to move some code out of GPUMaterializeEncoding.
