+++ compiler/src/iree/compiler/Codegen/Common/GPU/BUILD.bazel
@@ -75,6 +75,7 @@ iree_compiler_cc_library(
         "GPUTensorTileToSerialLoops.cpp",
         "GPUTile.cpp",
         "GPUTileReduction.cpp",
+        "GPUTileSwizzleUtils.cpp",
         "GPUVectorAlloc.cpp",
         "GPUVectorDistribution.cpp",
         "GPUVerifyDistribution.cpp",
